Fix metalint; move to /api/v1
