+++ src/coordinator/api/v1/handler/namespace/delete.go
@@ -28,7 +28,7 @@ import (
 	"strings"
 
 	clusterclient "github.com/m3db/m3cluster/client"
-	"github.com/m3db/m3db/src/coordinator/handler"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
 	"github.com/m3db/m3db/src/dbnode/storage/namespace"
