+++ src/coordinator/api/v1/handler/placement/delete_all_test.go
