+++ src/coordinator/api/v1/handler/placement/delete_test.go
