+++ src/coordinator/api/v1/handler/prometheus/native/read.go
@@ -27,8 +27,8 @@ import (
 	"net/http"
 
 	"github.com/m3db/m3db/src/coordinator/executor"
-	"github.com/m3db/m3db/src/coordinator/handler"
-	"github.com/m3db/m3db/src/coordinator/handler/prometheus"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler/prometheus"
 	"github.com/m3db/m3db/src/coordinator/parser/promql"
 	"github.com/m3db/m3db/src/coordinator/ts"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
