+++ src/coordinator/api/v1/handler/prometheus/native/read_test.go
@@ -30,7 +30,7 @@ import (
 	"time"
 
 	"github.com/m3db/m3db/src/coordinator/executor"
-	"github.com/m3db/m3db/src/coordinator/handler/prometheus"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler/prometheus"
 	"github.com/m3db/m3db/src/coordinator/test/local"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
