+++ src/coordinator/api/v1/handler/prometheus/remote/read.go
@@ -27,8 +27,8 @@ import (
 
 	"github.com/m3db/m3db/src/coordinator/executor"
 	"github.com/m3db/m3db/src/coordinator/generated/proto/prompb"
-	"github.com/m3db/m3db/src/coordinator/handler"
-	"github.com/m3db/m3db/src/coordinator/handler/prometheus"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler/prometheus"
 	"github.com/m3db/m3db/src/coordinator/storage"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
