+++ src/coordinator/api/v1/handler/prometheus/remote/read_test.go
@@ -33,7 +33,7 @@ import (
 
 	"github.com/m3db/m3db/src/coordinator/executor"
 	"github.com/m3db/m3db/src/coordinator/generated/proto/prompb"
-	"github.com/m3db/m3db/src/coordinator/handler/prometheus"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler/prometheus"
 	"github.com/m3db/m3db/src/coordinator/storage"
 	"github.com/m3db/m3db/src/coordinator/test"
 	"github.com/m3db/m3db/src/coordinator/test/local"
