+++ src/coordinator/api/v1/handler/prometheus/remote/write.go
@@ -25,8 +25,8 @@ import (
 	"net/http"
 
 	"github.com/m3db/m3db/src/coordinator/generated/proto/prompb"
-	"github.com/m3db/m3db/src/coordinator/handler"
-	"github.com/m3db/m3db/src/coordinator/handler/prometheus"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler/prometheus"
 	"github.com/m3db/m3db/src/coordinator/storage"
 	"github.com/m3db/m3db/src/coordinator/util/execution"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
