+++ src/coordinator/api/v1/httpd/handler_test.go
@@ -27,8 +27,8 @@ import (
 
 	"github.com/m3db/m3db/src/cmd/services/m3coordinator/config"
 	"github.com/m3db/m3db/src/coordinator/executor"
-	"github.com/m3db/m3db/src/coordinator/handler/prometheus/native"
-	"github.com/m3db/m3db/src/coordinator/handler/prometheus/remote"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler/prometheus/native"
+	"github.com/m3db/m3db/src/coordinator/api/v1/handler/prometheus/remote"
 	"github.com/m3db/m3db/src/coordinator/test/local"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
