+++ src/coordinator/services/m3coordinator/server/server.go
@@ -35,7 +35,7 @@ import (
 	"github.com/m3db/m3db/src/cmd/services/m3coordinator/config"
 	m3dbcluster "github.com/m3db/m3db/src/coordinator/cluster/m3db"
 	"github.com/m3db/m3db/src/coordinator/executor"
-	"github.com/m3db/m3db/src/coordinator/httpd"
+	"github.com/m3db/m3db/src/coordinator/api/v1/httpd"
 	"github.com/m3db/m3db/src/coordinator/policy/filter"
 	"github.com/m3db/m3db/src/coordinator/storage"
 	"github.com/m3db/m3db/src/coordinator/storage/fanout"
